+++ CHANGELOG.md
@@ -2,6 +2,12 @@ Changelog for Paco
 ==================
 
 
+9.1.1 (unreleased)
+------------------
+
+- Nothing changed yet.
+
+
 9.1.0 (2020-12-23)
 ------------------